% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk_str_filter.R
\name{awk_str_filter}
\alias{awk_str_filter}
\title{Filter arbitrary rows from a tab-delimited text file using awk}
\usage{
awk_str_filter(filename, filter_col, filter_str)
}
\arguments{
\item{filename}{The path to the tab-delimited text file.}

\item{filter_col}{The name of the column to filter.}

\item{filter_str}{A character vector of strings to filter for in the specified column.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing the filtered rows from the input file
}
\description{
This function wraps the system command \code{awk} to efficiently filter rows from a tab-delimited text file. The function accepts the path to a tab-delimited text file, name of a column to filter, and the string to filter for. The function returns a tibble containing the filtered rows.
}
\examples{
ldlr_variants <- awk_str_filter(filename = "/project/PMBB/PMBB-Release-2020-2.0/Exome/Variant_annotations/PMBB-Release-2020-2.0_genetic_exome_variant-annotation-counts.txt",
               filter_col = "Gene.refGene",
               filter_str = c("LDLR"))
}
