% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmbb_extract_genotype_masks.R
\name{pmbb_extract_genotype_masks}
\alias{pmbb_extract_genotype_masks}
\title{Extract genotypes based on variant masks}
\usage{
pmbb_extract_genotype_masks(
  gene,
  annotation_file,
  gene_col,
  masks,
  variant_id_col,
  effect_allele_col,
  mask_operator = NULL,
  plink_bin,
  bfile,
  threads = 1,
  memory = 8000
)
}
\arguments{
\item{gene}{The name of the gene to filter variants for.}

\item{annotation_file}{The path to the variant annotation file.}

\item{gene_col}{The name of the column in the annotation file to filter the gene on.}

\item{masks}{A named list of lists containing filter criteria for each mask.
Each element of the outer list should be named after the desired output name for the filtered data frame (eg. name of the variant mask).
Each inner list should contain the filter criteria for a specific set of columns.
For example: list(plof_0.001 = list(ExonicFunc.ensGene = "== 'stopgain'", gnomAD_exome_ALL = "< 0.001"), common_0.01 = list(gnomAD_exome_ALL = "> 0.01"))}

\item{variant_id_col}{The name of the column in the annotation file that contains the variant IDs.}

\item{effect_allele_col}{The name of the column in the annotation file that contains the effect alleles.}

\item{mask_operator}{A named list specifying the operation to perform for each mask. Valid operations are "burden" (default) and "single".
The names in the mask_operator list should correspond to the names in the masks list.
For example: list(plof_0.001 = "burden", common_0.01 = "single")}

\item{plink_bin}{The path to the \code{plink2} binary executable.}

\item{bfile}{The prefix of the plink binary files (without the file extension).}

\item{threads}{The number of threads to use for parallel processing (default: 1).}

\item{memory}{The amount of memory (in MB) to allocate for \code{plink2} (default: 8000).}
}
\value{
A named list, where each element represents a mask and contains a list with components based on the specified mask_operator:
- "variants": A \code{\link[tibble:tibble]{tibble::tibble()}} of the variants included in the mask.
- "genotypes": A \code{\link[tibble:tibble]{tibble::tibble()}} containing the genotype for each PMBB participant.
- "mask": A string containing the mask applied to filter the variants.
- "mask_type": A \code{\link[tibble:tibble]{tibble::tibble()}} containing the mask type applied (eg. "burden" or "single").
}
\description{
This function extracts genotypes for a subset of participants based on a set of masks. It wraps \code{awk_str_filter}, \code{apply_variant_masks}, and \code{pmbb_extract_genotypes} to extract genotypes for PMBB participants based on a set of masks.
}
\examples{
\dontrun{
cftr_mask_res <- pmbb_extract_genotype_masks(
  gene = "CFTR",
  annotation_file = "/project/PMBB/PMBB-Release-2020-2.0/Exome/Variant_annotations/PMBB-Release-2020-2.0_genetic_exome_variant-annotation-counts.txt",
  gene_col = "Gene.refGene",
  masks = list(
    plof_lt_0.001 = list(ExonicFunc.ensGene = "\%in\% c('stopgain', 'stoploss', 'frameshift substitution')", gnomAD_exome_ALL = "< 0.001"),
    p_lp_het_gt_10 = list(CLNSIG = "\%in\% c('Pathogenic', 'Pathogenic/Likely_pathogenic')", HET_REF_ALT_CTS = "> 10")
  ),
  mask_operator = list(
    plof_lt_0.001 = "burden",
    p_lp_het_gt_10 = "single"
  ),
  variant_id_col = ID,
  effect_allele_col = Alt,
  plink_bin = "/project/voltron/Applications/PLINK/plink2_linux_avx2_20230607/plink2",
  bfile = "/project/PMBB/PMBB-Release-2020-2.0/Exome/pVCF/all_variants/PMBB-Release-2020-2.0_genetic_exome_GL"
)
}
}
\seealso{
Other {genotypes}: 
\code{\link{pmbb_extract_genotypes}()}
}
\concept{{genotypes}}
