% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmbb_extract_genotypes.R
\name{pmbb_extract_genotypes}
\alias{pmbb_extract_genotypes}
\title{Extract genotypes from PMBB plink files}
\usage{
pmbb_extract_genotypes(
  variant_df,
  variant_id_col,
  effect_allele_col,
  plink_bin,
  bfile,
  threads = 1,
  memory = 8000
)
}
\arguments{
\item{variant_df}{A data frame containing the variant information.}

\item{variant_id_col}{The name of the column in \code{variant_df} that contains the variant IDs.}

\item{effect_allele_col}{The name of the column in \code{variant_df} that contains the effect alleles.}

\item{plink_bin}{The path to the \code{plink2} binary executable.}

\item{bfile}{The prefix of the plink binary files (without the file extension).}

\item{threads}{The number of threads to use for parallel processing (default: 1).}

\item{memory}{The amount of memory (in MB) to allocate for \code{plink2} (default: 8000).}
}
\value{
A long-format \code{\link[tibble:tibble]{tibble::tibble()}} containing the extracted genotypes for each PMBB participant.
}
\description{
This function is a wrapper around \code{plink2} to extract genotypes from PMBB plink files. The function accepts the path to the plink files and the list of variant IDs to extract. The function returns a long-format \code{\link[tibble:tibble]{tibble::tibble()}} containing the extracted genotypes for each PMBB participant.
}
\examples{
\dontrun{
pmbb_extract_genotypes(
  variant_df = cftr_variants,
  variant_id_col = ID,
  effect_allele_col = Alt,
  plink_bin = "/project/voltron/Applications/PLINK/plink2_linux_avx2_20230607/plink2",
  bfile = "/project/PMBB/PMBB-Release-2020-2.0/Exome/pVCF/all_variants/PMBB-Release-2020-2.0_genetic_exome_GL"
)
}
}
\seealso{
Other {genotypes}: 
\code{\link{pmbb_extract_genotype_masks}()}
}
\concept{{genotypes}}
