% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmbb_filter_exome_annotations_gene.R
\name{pmbb_filter_exome_annotations_gene}
\alias{pmbb_filter_exome_annotations_gene}
\title{Filter the PMBB whole exome sequence annotations file}
\usage{
pmbb_filter_exome_annotations_gene(annotations_filename, gene_names)
}
\arguments{
\item{annotations_filename}{Path to the PMBB whole-exome sequencing variant annotations file}

\item{gene_names}{A character vector of gene names used to filter the variant annotations file}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing the filtered variant annotations
}
\description{
This function filters the variant annotations file to only include variants within the specified gene. Because the variant annotation file is extremely large, this function wraps the \code{awk} shell command
}
\examples{
pmbb_filter_exome_annotations_gene(annotations_filename = "/project/PMBB/PMBB-Release-2020-2.0/Exome/Variant_annotations/PMBB-Release-2020-2.0_genetic_exome_variant-annotation-counts.txt",
                              gene_names = c("LDLR"))
}
