% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmbb_format_phecodes.R
\name{pmbb_format_labs}
\alias{pmbb_format_labs}
\title{Format labs data for PheWAS}
\usage{
pmbb_format_labs(labs_files, patient_class = "Outpatient")
}
\arguments{
\item{labs_files}{A character vector of file paths to the labs data}

\item{patient_class}{A character vector of patient classes (location of lab testing) to include in the analysis. Default is \code{"Outpatient"}}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing \code{PMBB_ID} and summary measures of lab values for each PMBB participant
}
\description{
This function will read in labs data from PMBB and format it for use in a PheWAS. By default, this function will return a data frame with summary measures (\code{min}, \code{median}, \code{mean}, \code{max}) of each lab value for each individual in the PMBB cohort. Because PMBB lab data is only roughly cleaned, the function will by default select only the most common \code{RESULT_NAME} for each lab test.
}
\examples{
\dontrun{
pmbb_labs_df <- fs::dir_ls("/project/PMBB/PMBB-Release-2020-2.0/Phenotype/2.3/", glob = "*labs*", recurse = TRUE) \%>\%
  pmbb_format_labs()
}
}
\seealso{
Other {phenotypes}: 
\code{\link{pmbb_format_phecodes}()}
}
\concept{{phenotypes}}
