% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmbb_format_phecodes.R
\name{pmbb_format_phecodes}
\alias{pmbb_format_phecodes}
\title{Format PMBB phecode data}
\usage{
pmbb_format_phecodes(phecode_file, phecode_case_threshold = 2)
}
\arguments{
\item{phecode_file}{Path to PMBB phecode file used to run PheWAS}

\item{phecode_case_threshold}{Number of cases required to classify an individual as a case}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing \code{PMBB_ID} and phecodes as columns, with TRUE/FALSE values for each phecode column
}
\description{
Given a dataframe containing PMBB phecodes, apply a case treshold for categorizing individuals as cases or controls
}
\examples{
\dontrun{
pmbb_phecodes_df <- pmbb_format_phecodes(phecode_file = "/project/PMBB/PMBB-Release-2020-2.0/Phenotype/2.3/PMBB-Release-2020-2.3_phenotype_PheCode-matrix.txt", phecode_case_threshold = 2)
}
}
\seealso{
Other {phenotypes}: 
\code{\link{pmbb_format_labs}()}
}
\concept{{phenotypes}}
