% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmbbPheWASr-package.R
\docType{package}
\name{pmbbPheWASr-package}
\alias{pmbbPheWASr}
\alias{pmbbPheWASr-package}
\title{pmbbPheWASr: Penn Medicine Biobank Phenome-wide Association Studies}
\description{
A Set of tools to perform phenome-wide association studies using Penn Medicine Biobank data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mglev1n/pmbbPheWASr}
  \item Report bugs at \url{https://github.com/mglev1n/pmbbPheWASr/issues}
}

}
\author{
\strong{Maintainer}: Michael Levin \email{michael.levin@pennmedicine.upenn.edu} (\href{https://orcid.org/0000-0002-9937-9932}{ORCID})

}
\keyword{internal}
