% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_pmbb_phewas_report.R
\name{render_pmbb_phewas_report}
\alias{render_pmbb_phewas_report}
\title{Render PheWAS Report}
\usage{
render_pmbb_phewas_report(mask_output_path, phewas_output_path, output_file)
}
\arguments{
\item{mask_output_path}{Path to the saved genotype output for each mask}

\item{phewas_output_path}{Path to the saved PheWAS output dataframe}

\item{output_file}{Full path to save the HTML report}
}
\value{
Path to the HTML report
}
\description{
This function renders an HTML report from pre-saved genotype mask and PheWAS output files.
}
\examples{
\dontrun{
render_pmbb_phewas_report()
}
}
